Wczytywanie planszy

+++ Menu.h
@@ -1,2 +1,2 @@
-int nowaPlansza(unsigned short int nY, unsigned short int nX);
-int Menu(unsigned short int nY, unsigned short int nX);
+int nowaPlansza(int nY, int nX);
+int Menu(int nY, int nX);
